requirements.txt
# YOLO11 + SAM2 项目依赖

# 核心深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# YOLO相关
ultralytics>=8.0.0

# 图像处理
opencv-python>=4.8.0
Pillow>=9.0.0
matplotlib>=3.5.0
numpy>=1.21.0

# 进度条和工具
tqdm>=4.64.0
psutil>=5.9.0

# 配置文件处理
PyYAML>=6.0

# SAM2相关（需要从GitHub安装）
# git+https://github.com/facebookresearch/segment-anything-2.git

# 可选：Jupyter支持
# jupyter>=1.0.0
# ipywidgets>=8.0.0